Fix various small warnings
This commit addresses the following warnings:
- Remove unnecessary semi-colons.
- Use `isEmpty()` instead of comparing the size to 0 (for collections) or the value to `""` (for strings).
- Remove unnecessary calls to `toString()`.
- Remove redundant type declarations.
- Use Kotlin's property instead of getter/setter.
- Replace redundant `StringBuilder` with `String`.
- Remove unnecessary casts.

+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/util/SntpClient.java
@@ -269,7 +269,7 @@ public final class SntpClient {
         // Extract the results.
         final byte leap = (byte) ((buffer[0] >> 6) & 0x3);
         final byte mode = (byte) (buffer[0] & 0x7);
-        final int stratum = (int) (buffer[1] & 0xff);
+        final int stratum = buffer[1] & 0xff;
         final long originateTime = readTimestamp(buffer, ORIGINATE_TIME_OFFSET);
         final long receiveTime = readTimestamp(buffer, RECEIVE_TIME_OFFSET);
         final long transmitTime = readTimestamp(buffer, TRANSMIT_TIME_OFFSET);
